Updated models ; Added missing endpoints ; Added beta voucher properties ; Upgraded dependencies ; Bumping version to 4.5.0

+++ src/main/java/be/woutschoovaerts/mollie/data/invoice/InvoiceLine.java
@@ -18,7 +18,7 @@ public class InvoiceLine {
 
     private int count;
 
-    private double vatPercentage;
+    private int vatPercentage;
 
     private Amount amount;
 }
